Renamed `gui` dir to `app`
