+++ src/app/assets/white_king.png
